# experimental: remove return value of FunctionListener.Before