Another follow up from our discussion in #wazero-dev, this PR modifies the signature of `FunctionListener.Before` to remove the returned `context.Context`.

This change aims to remove the internal context stack tracking return values of the `Before` methods. For context, I am copying a quote from the Slack discussion:

> While I find the API somewhat elegant, it has proven to be unusable in most cases where we've used FunctionListener because embedding values in a context requires at least one heap allocation, but sometimes two or more. The listeners often being on hot code paths of the guest program, and this heap allocation ends up being extremely costly.
>
> We worked around the issue by implementing internal stacks within the listeners, in the same way, that the wazero engine keeps track of a stack of the context.Context returned by Before. Implementing those stacks as slices allows reusing the underlying memory, which ends up being an extremely cheap mechanism for correlation between the Before and After methods: the allocation of context values added 50 to 100ns of latency per function, while the internal stack ends up adding 5 to 10ns (~10x cheaper).
>
> However, we still end up paying for the engine's internal context stack, despite not needing it. On performance-critical code paths, it almost doubles the overall latency of invoking function listeners.

## Using stacks to correlate Before and After

One of the issues was the fact that `After` was not always being called. For example, if a module hit a trap, or a host function panicked, the call to the `After` method would be skipped, which made it challenging to implement stack-based context tracking internally in the listener. This limitation actually impacted the host logging listeners and caused multiple tests within wazero to fail after making the initial change.

I am suggesting that we address it by adding a third `Abort` method on `FunctionListener`, which is invoked when a module execution is aborted, in which case there are no result values to pass to `After` and instead of communicate the reason for aborting to the call to `Abort`.

The guarantee that either `After` or `Abort` will be called allows listeners to use an internal stack to carry state from `Before` to one of these methods.

## State propagation through the call stack

One of the details that @codefromthecrypt shared in #wazero-dev was that the second reason for returning a `context.Context` was to pass state not just between `Before` and `After`, but also to other listeners and host function calls through the call stack. Use cases for that include telemetry like tracing, where a listener creates a span and injects it in the context that eventually gets passed as input to another callee.

While this use case isn't covered in the change I'm submitting here, I believe that it can be served by injecting a _mutable context value_ (as @codefromthecrypt described it). Here would be an example:

```go
// Completely hypothetical API, just to demonstrate the idea

rootSpan := &tracing.Span{
  ParentSpan: ...,
}

ctx := context.WithValue(ctx, SpanKey{}, &rootSpan)
...
// eventually instantiate module or call exported function with ctx
```
```go
func (f *functionTracer) Before(ctx context.Context, ...) {
  spanPtr := ctx.Value(SpanKey{}).(**tracing.Span)
  span := &tracing.Span{
    ParentSpan: *spanPtr,
    ...
  }
  *spanPtr = span // push
  span.Start()
}

func (f *functionTracer) After(ctx context.Context, ...) {
  spanPtr := ctx.Value(SpanKey{}).(**tracing.Span)
  span := *spanPtr
  span.Finish()
  *spanPtr = spanPtr.Parent // pop
}

// + Abort
```
In the future, injecting and extracting the context value may also be abstracted into a generic mechanism to bridge between the `FunctionListener` interface and higher level abstractions.

Some of the benefits I'm seeing to this approach is that it reduces the baseline cost of function listeners, the higher cost of maintaining stacks is only paid when necessary, and can be optimized individually by each listener to suite the use cases.

The trade off is that we offer something that is a bit less _packaged_, in some ways we are lowering the abstraction level of function listeners, leaving more of the implementation decisions to the host.

Please take a look and let me know if you would like to see anything changed!