@achille-roussel 2023-05-11
> I can see the sense in not having the error param, and also the return removal, just not adding Abort when it isn't called (unless I am missing that). I think if we are removing things because they aren't implemented, we shouldn't add new unimplemented things as it makes it confusing. Can you separate these steps?

This was probably a miss on my part in the description, I can try to clarify.

The host logging code was depending on context propagation to calculate things like indentation levels. Removing the context means we needed a different way. Because the intention was to move state tracking into the function listener instead of having it in the wazero engine, I modified the loggers to track their state without the context, which highlighted the fact that sometimes `After` is not called.

For example, `Test_procExit` makes two calls to `proc_exit` which never calls `After` because the function never returns, and invalidated the state, causing the output to change and the tests to fail. This is just one example of the many places where we use the loggers to validate beahvior, wasi_snapshot_preview1 had many such tests.

In order to fix the issue and reconcile the state of the loggers, I added the `Abort` logic so everything that gets pushed to the inner stack gets popped. Now I couldn't just ensure that `After` was going to always be called because that would cause the loggers to print return values in places where the tests validated that there would be no returns (e.g. `proc_exit`).

I agree that adding `Abort` is expanding the scope, here is the place where it's needed in wazero https://github.com/tetratelabs/wazero/pull/1453/files#diff-d60dc448f4c159bc82ccabf4dca5fb109524074e4fc7810d766bcecb7a649af4R229-R232

Something I have to admit I did not explore was to take the approach I described in **State propagation through the call stack** with the loggers. I can probably get around not adding `Abort` with this model.

After exploring this code further, I feel like the presence of `Abort` is meaningful, tho I understand the desire not to couple changes:
- I can rework the loggers to not need `Abort`, and submit that as a separate PR if you think it's important to separate the change and discussion
- we can continue the discussion here until we feel comfortable with the change or choose not to, there is no rush!