@achille-roussel 2023-05-11
> A test may show what I mean, for example, if there's a call stack of 3 I would expect not all three listeners to get the Abort(). If I'm wrong, please correct me! In any case, a test maybe can explain this working or not working.

My intention is for each call to `Before` on a listener to be matched with either one call to `After` or one call to `Abort`. I believe that's what I implemented because some of the existing tests wouldn't pass otherwise, but I can add an extra test to explicitly validate that.

To illustrate, say there is a call stack like this, each function have listeners installed:
```
f1
. f2
  . f3
```
When `f1`, then `f2`, and `f3` are called, each of their listeners' `Before` method is invoked.

- If each of these functions returns normally, their listeners `After` methods are called, but `Abort` is not
- If `f3` traps, then `Abort` is called on the listener of `f3`, `f2`, and `f1`, none of their `After` methods are called
- If `f3` returns, but after that `f2` traps, the `After` method of `f3`'s listener is called, but the `Abort` method of `f2`, then `f1` are called

It's the guarantee that for each call to `Before` exactly one of `After` or `Abort` will be called which allows listeners to track state associated with the call stack internally, and it doesn't get out of sync.